requirements.txt
addict==2.4.0
attributee==0.1.5
certifi==2021.5.30
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cycler==0.11.0
Cython==0.29.28
dataclasses==0.8
decord==0.6.0
dotty-dict==1.3.0
einops==0.4.1
Flask==2.0.3
flatbuffers==2.0
fonttools==4.25.0
idna==3.7
imgviz==1.6.2
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.1
labelme==5.1.1
lap==0.4.0
llvmlite==0.36.0
MarkupSafe==2.0.1
matplotlib==3.3.4
mmcls==0.22.1
mmcv==1.4.8
mmdet==2.23.0
motmetrics==1.2.5
natsort==8.2.0
numba==0.53.0
numpy==1.19.5
onnx==1.7.0
onnxruntime-gpu==1.10.0
opencv-contrib-python==4.6.0.66
opencv-python==4.5.5.64
packaging==21.3
pandas==1.1.5
Pillow==8.4.0
protobuf==3.19.4
pycocotools==2.0.2
pyparsing==3.0.8
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
QtPy==2.0.1
regex==2022.4.24
requests==2.27.1
scipy==1.5.4
seaborn==0.11.2
setuptools-scm==6.4.2
six==1.16.0
termcolor==1.1.0
terminaltables==3.1.10
timm==0.5.4
tomli==1.2.3
tqdm==4.64.0
typing_extensions==4.1.1
urllib3==1.26.18
Werkzeug==2.0.3
wincertstore==0.2
xmltodict==0.12.0
yapf==0.32.0
zipp==3.6.0
